requirements.txt
Flask
Pillow
gunicorn
requests
beautifulsoup4
pymongo[srv]
python-dotenv
certifi